+++ src/components/SongCard.tsx
@@ -1,6 +1,6 @@
 import { Button, makeStyles } from '@material-ui/core';
 import React from 'react';
-import { SongInfo } from '../types/SongInfo';
+import { SongInfo } from '../types/songInfo';
 
 const useStyles = makeStyles({
   root: {
