+++ src/types/api.ts
@@ -1,0 +1,7 @@
+// Error response after extracting useful attributes
+export interface ApiErrorResponse {
+  code: number;
+  body?: any & {
+    error?: string;
+  };
+}
