+++ src/types/playerInfo.ts
