+++ src/types/playingNote.ts
