+++ src/types/song.ts
