+++ src/types/songInfo.ts
